Refactored backend

+++ backend/models/course.py
@@ -11,7 +11,7 @@ class Course:
     course_number: int
     long_title: Optional[str] = None
     offered_terms: Set[str] = field(default_factory=set)
-    credit_hours: Optional[int] = None
+    credit_hours: Optional[float] = None
     distribution: Optional[str] = None
     analyzing_diversity: bool = False
     cross_list: List[str] = field(default_factory=list)
